% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nass_data.R
\name{nass_data}
\alias{nass_data}
\title{Get data from the USDA NASS data query.}
\usage{
nass_data(source_desc = NULL, sector_desc = NULL, group_desc = NULL,
  commodity_desc = NULL, class_desc = NULL, prodn_practice_desc = NULL,
  util_practice_desc = NULL, statisticcat_desc = NULL, unit_desc = NULL,
  short_desc = NULL, domain_desc = NULL, domaincat_desc = NULL,
  agg_level_desc = NULL, state_ansi = NULL, state_fips_code = NULL,
  state_alpha = NULL, state_name = NULL, asd_code = NULL,
  asd_desc = NULL, county_ansi = NULL, county_code = NULL,
  county_name = NULL, region_desc = NULL, zip_5 = NULL,
  watershed_code = NULL, watershed_desc = NULL,
  congr_district_code = NULL, country_code = NULL, country_name = NULL,
  location_desc = NULL, year = NULL, freq_desc = NULL,
  begin_code = NULL, end_code = NULL, reference_period_desc = NULL,
  week_ending = NULL, load_time = NULL, token = NULL, ...)
}
\arguments{
\item{source_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{sector_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{group_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{commodity_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{class_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{prodn_practice_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{util_practice_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{statisticcat_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{unit_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{short_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{domain_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{domaincat_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{agg_level_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{state_ansi}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{state_fips_code}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{state_alpha}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{state_name}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{asd_code}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{asd_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{county_ansi}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{county_code}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{county_name}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{region_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{zip_5}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{watershed_code}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{watershed_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{congr_district_code}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{country_code}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{country_name}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{location_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{year}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{freq_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{begin_code}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{end_code}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{reference_period_desc}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{week_ending}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{load_time}{See USDA webpage for explanations. You can use nass_param() to see the values.}

\item{token}{API key}
}
\value{
A data frame
}
\description{
Limit for data request would not return more than 50,000 records/rows.
It is better to check count of records using nass_count()
}

