% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nass_param.R
\name{nass_param}
\alias{nass_param}
\title{Get metadata about USDA NASS column/header names.}
\usage{
nass_param(param = NULL, token = NULL)
}
\arguments{
\item{param}{A valid parameter value}

\item{token}{API key}
}
\value{
A list containing metadata as data frame.
}
\description{
Available names are:
source_desc, sector_desc, group_desc, commodity_desc,
class_desc, prodn_practice_desc, util_practice_desc, statisticcat_desc,
unit_desc, short_desc, domain_desc, domaincat_desc, agg_level_desc,
state_ansi, state_fips_code, state_alpha, state_name, asd_code,
asd_desc, county_ansi, county_code, county_name, region_desc, zip_5,
watershed_code, watershed_desc, congr_district_code, country_code,
country_name, location_desc, year, freq_desc, begin_code, end_code,
reference_period_desc, week_ending, load_time, value, and CV %
}

